% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSecondaryMarket.R
\name{GetSecondaryMarket}
\alias{GetSecondaryMarket}
\title{Download loans in secondary market}
\usage{
GetSecondaryMarket()
}
\description{
Download the loans currently listed on the secondary market
}

