% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClubModel.R
\docType{package}
\name{LendingClubModel}
\alias{LendingClubModel}
\alias{LendingClubModel-package}
\title{Lending Club Model}
\description{
To document the process of building a model which increasing
investor returns.
}
\seealso{
\itemize{
\item{\url{http://ryankuhn.net/LendingClubModel}}
\item{\url{http://github.com/kuhnrl30/LendingClubModel/issues} for bug reports}
}
}
\author{
Ryan Kuhn, CPA, \email{kuhnrl30@gmail.com}
}
