% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P2PLending.R
\docType{data}
\name{LoanData}
\alias{LoanData}
\title{Loan Data}
\usage{
data(LoanData)
vignette("GetTheData")
}
\description{
Historical loan data downloaded from Lending Club and processed in
vignette("GetTheData")
}

