% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenSession.R
\name{OpenSession}
\alias{OpenSession}
\title{Open a session with Lending Club}
\usage{
OpenSession(email, password)
}
\arguments{
\item{email}{login information for Lending Club}

\item{password}{login information for Lending Club}
}
\description{
Open a session with Lending Club
}

