% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P2PLending.R
\docType{package}
\name{P2PLending}
\alias{P2PLending}
\alias{P2PLending-package}
\title{P2P Lending}
\usage{
vignette("ExploratoryAnalysis")
}
\description{
To create the tools for working with the Lending Club API and
to document the process to build a model for increasing
investor returns.
}
\author{
Ryan Kuhn, CPA
}

