% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LendingClubModel.R
\docType{package}
\name{LendingClubModel}
\alias{LendingClubModel}
\alias{LendingClubModel-package}
\title{Lending Club Model}
\description{
To document the process to build a model for increasing
investor returns.
}
\author{
Ryan Kuhn, CPA
}
